+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `var`, `std`, `take`, `count`, and `length` to vectors
 - Added `sample` to data frames
+- Improved error message for `read_csv` with no headers
 
 ## 0.1.0 (2020-05-13)